requirements.txt
altair==5.5.0
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.10.0
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
filelock==3.18.0
fonttools==4.59.0
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hf-xet==1.1.7
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib-metadata==6.11.0
Jinja2==3.1.6
joblib==1.3.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.10.1
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
mdurl==0.1.2
msgpack==1.1.1
narwhals==2.1.0
numba==0.61.2
numpy==1.26.4
packaging==23.2
pandas==2.0.3
Pillow==10.1.0
platformdirs==4.3.8
plotly==5.18.0
pooch==1.8.2
protobuf==4.25.8
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.31.0
rich==13.9.4
rpds-py==0.27.0
scikit-learn==1.3.2
scipy==1.11.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soundfile==0.12.1
soxr==0.5.0.post1
streamlit==1.29.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.1.0
validators==0.35.0
watchdog==6.0.0
zipp==3.23.0
